requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
djangorestframework==3.12.4
drf-nested-routers==0.93.3
importlib-metadata==4.6.3
iniconfig==1.1.1
kombu==5.1.0
packaging==21.0
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-slugify==5.0.2
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
